Move blob storage calls out of ServerApi for #2987
ServerApi is the client for our API. The upload PUT goes to blob storage, which
is a different server sharing none of its conventions: no bearer token, no JSON,
not our base URL, no 401-refresh flow. Two things gave that away in the code -
the method needed a comment explaining why it does not use executeRequest, and
it took an absolute URL where every other method takes a path relative to our
API base.

It now lives in a fileStorage module, together with uploadFile, which spans both
servers and belongs with neither: our API reserves the file and issues a signed
URL, then the bytes go straight to storage. uploadFile also moves out of the
scraps folder, since nothing about it is scrap-specific and the editor will want
it for inline images.

getContentType is shared rather than duplicated so that the type we record when
reserving the upload and the type we store on the blob cannot drift apart.

ServerApi keeps createFileUpload and getFileUrl - those really are our API.
